+++ validator.go
@@ -2,6 +2,7 @@ package main
 
 import (
 	"fmt"
+	"sort"
 	"strings"
 )
 
@@ -13,6 +14,29 @@ func isZeroAmount(amounts map[string]Amount) bool {
 	}
 	return true
 }
+
+func buildImbalancedTransactionMsg(
+	amounts map[string]Amount,
+) string {
+	currencies := make([]string, 0, len(amounts))
+	for currency := range amounts {
+		currencies = append(currencies, currency)
+	}
+	sort.Strings(currencies)
+
+	amountStrs := []string{}
+	for _, currency := range currencies {
+		amount := amounts[currency]
+		amountAndCurrency := fmt.Sprintf("%v %v", amount, currency)
+		amountStrs = append(amountStrs, amountAndCurrency)
+	}
+	msg := fmt.Sprintf(
+		"imbalanced transaction, (total amount) = %v",
+		strings.ReplaceAll(strings.Join(amountStrs, " + "), "+ -", "- "),
+	)
+	return msg
+}
+
 func newValidator(filePath, accountsPath string) *Validator {
 	validator := Validator{
 		filePath:     filePath,
